Adding arguments to work around Gitlab
The URLDownload argument is needed so that the recipe creates the downloads folders on the initial run. The following work arounds fail without this in here for the first run.
The FileCreator creates a placeholder to work around GitLab's tendency to return the same etag across multiple URLDownloader runs.
The PathDeleter clears out the old download so that we force a new download from GitLab.