Update for sh-compatibility
Tested and made sure the script works in sh as is. Quoted all variables. And referenced full path to each binary unless it was a built-in (e.g. echo).

There are just two changes that need to be made to make it fully sh-compatible. There are two if statements that make use of regex comparisons inside double brackets on line 47 and line 91. May need to come back to this at another time to finish.