Update AudioHijack.munki to include ACE install
Audio Hijack requires the Rogue Amoeba ACE.driver Audio plugin. If it is not installed or current on launch, the user is presented with a GUI to install it - requiring admin rights.

There is a compiled binary at `Audio Hijack.app/Contents/Resources/aceinstaller`. The usage for that binary is:

```
aceinstaller: <action>
    help
        Display this message.
    install [-s|--silent]
        Install (or update) the ACE.driver coreaudiod plugin. The version to be installed must be
        newer than what is currently installed, and the current system version must be supported.
        Installing requires root privileges.
        -s|--silent    Do not generate an error if the version to be installed is not newer than
                       than the version that is currently installed.
    uninstall [-s|--silent]
        Uninstall the ACE.driver coreaudiod plugin. There must be a plugin installed for it to be
        uninstalled. Uninstalling requires root privileges.
    version
        Display the installed and installable ACE versions.
```

I've tested using the command `aceinstaller install -s` in a pkginfo postinstall script on 10.14-11.2.3 and all installed the latest version.
When updating Audio Hijack from 3.8.3 to 3.8.4, there was a GUI prompt that said the ACE driver needed to be updated, so this postinstall_script should run every time Audio Hijack is installed to make sure it is using the latest, current ACE.driver.