Enhance JamfSetupManager.download
This adds support for downloading (or choosing not to download) the pre-release version if it is the most recent release. As well, since the pkg being downloaded generates a reasonable filename for most uses, renaming the pkg is removed from the download recipe. If the user wants the %NAME%-%version%.pkg name, they can run the pkg recipe, since it renames the pkg to that. If they prefer the version number with build number that automatically comes with the GitHub-provided pkg, they can use the download recipe as a parent.